backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
Pillow>=9.0.0
joblib==1.3.2
scikit-learn>=1.2.0
scikit-image>=0.25.0
numpy>=1.23.0
pandas>=1.5.0
matplotlib==3.8.2
tensorflow>=2.10.0
tensorflow-hub>=0.12.0
gunicorn==21.2.0
tensorflow>=2.15.0
Pillow>=10.0.0
torch>=2.1.0